Update other mods on settings update
