+++ gui/settings.py
@@ -11,6 +11,8 @@ from PySide6.QtWidgets import (
     QWidget,
     QAbstractButton,
     QInputDialog,
+    QSpacerItem,
+    QSizePolicy,
 )
 
 import pyclip
@@ -26,7 +28,7 @@ from constants.itemconstants import (
     GRATITUDE_CRYSTAL,
     GROUP_OF_TADTONES,
 )
-from filepathconstants import BASE_PRESETS_PATH, CONFIG_PATH, ITEMS_PATH, PRESETS_PATH
+from filepathconstants import BASE_PRESETS_PATH, COMBINED_MODS_FOLDER, CONFIG_PATH, ITEMS_PATH, OTHER_MODS_PATH, PRESETS_PATH
 from gui.components.list_pair import ListPair
 from gui.components.tristate_check_box import RandoTriStateCheckBox
 from gui.mixed_entrance_pools import MixedEntrancePools
@@ -384,6 +386,8 @@ class Settings:
         if isinstance(from_widget, MixedEntrancePools) and widget_info is not None:
             mixed_entrance_pools = [pool for pool in widget_info if len(pool) > 0]
             self.config.settings[0].mixed_entrance_pools = mixed_entrance_pools
+
+        self.generate_other_mods_list()
 
         if allow_rewrite:
             write_config_to_file(CONFIG_PATH, self.config)
@@ -900,3 +904,51 @@ class Settings:
 
         seed_rng(self.config)
         self.ui.hash_label.setText(f"Hash: {self.config.get_hash()}")
+
+    def generate_other_mods_list(self):
+        self.main.clear_layout(self.ui.other_mods_scroll_layout)
+        found_mods = []
+        for mod_path in OTHER_MODS_PATH.glob("*"):
+            if mod_path.is_dir():
+                mod_name = mod_path.name
+
+                # Don't include the combined mods folder as a mod folder. Normally this folder is deleted, but
+                # if generation fails for some reason, then it might not get deleted.
+                if mod_name == COMBINED_MODS_FOLDER:
+                    continue
+
+                # Don't include the mod if it has an exefs folder. We don't support integrating other mods which modify code
+                if (OTHER_MODS_PATH / mod_name / "exefs").exists():
+                    continue
+
+                mod_checkbox = QCheckBox(mod_name)
+                mod_checkbox.clicked.connect(self.update_mods_in_config)
+                if mod_name in self.config.settings[0].other_mods:
+                    mod_checkbox.setChecked(True)
+                    found_mods.append(mod_name)
+
+                self.ui.other_mods_scroll_layout.addWidget(mod_checkbox)
+
+        # Add a vertical spacer to push the mod list up
+        self.ui.other_mods_scroll_layout.addSpacerItem(
+            QSpacerItem(
+                20, 40, QSizePolicy.Policy.Minimum, QSizePolicy.Policy.Expanding
+            )
+        )
+
+        # Remove mods from config which weren't found
+        for mod_name in self.config.settings[0].other_mods.copy():
+            if mod_name not in found_mods:
+                print(
+                    f"Removing mod {mod_name} from other_mods list as the QCheckbox for the mod could not be found"
+                )
+                self.config.settings[0].other_mods.remove(mod_name)
+
+    def update_mods_in_config(self):
+        other_mods = self.config.settings[0].other_mods
+        other_mods.clear()
+        for checkbox in self.ui.other_mods_scroll_widget.findChildren(QCheckBox):
+            if checkbox.isChecked():
+                other_mods.append(checkbox.text())
+
+        self.update_from_gui()
